fix all type errors

+++ botFunctions.py
@@ -1,8 +1,17 @@
+from typing import Dict
 from telegram.error import TelegramError
 from telegram import Update
 
+from dataStructures import InvalidActions
+
 
 async def delete_message(update: Update, reason: str):
+    if (
+        update.effective_user is None
+        or update.effective_chat is None
+        or update.message is None
+    ):
+        return
     try:
         await update.message.delete()
         await update.effective_chat.send_message(
@@ -14,12 +23,21 @@ async def delete_message(update: Update, reason: str):
         await update.effective_chat.send_message("כשל במחיקת הודעה: " + e.message)
 
 
-async def ban_user(update: Update):
+kick_messages: Dict[InvalidActions, str] = {
+    InvalidActions.none: "",
+    InvalidActions.swear: "לאחר שקילל והוזהר בפעם השלישית",
+    InvalidActions.arabic: "לאחר שרשם בערבית",
+}
+
+
+async def ban_user(update: Update, reason: InvalidActions):
+    if update.effective_user is None or update.effective_chat is None:
+        return
     try:
         if update.effective_user.id != 227093322:
             await update.effective_chat.ban_member(update.effective_user.id)
         await update.effective_chat.send_message(
-            f"המשתמש {update.effective_user.full_name} הועף מהקבוצה לאחר הפרת הכללים 3 פעמים..."
+            f"המשתמש {update.effective_user.full_name} הועף מהקבוצה {kick_messages[reason]}..."
         )
     except TelegramError as e:
         await update.effective_chat.send_message("כשל בזריקת משתמש: " + e.message)
